refractor: move function getPrincipalsFromRequest to a util function
