+++ src/plugins/workspace/server/saved_objects/workspace_saved_objects_client_wrapper.ts
@@ -28,11 +28,9 @@ import {
   SavedObjectsDeleteByWorkspaceOptions,
   SavedObjectsErrorHelpers,
 } from '../../../../core/server';
-import {
-  SavedObjectsPermissionControl,
-  SavedObjectsPermissionControlContract,
-} from '../permission_control/client';
+import { SavedObjectsPermissionControlContract } from '../permission_control/client';
 import { WorkspaceFindOptions } from '../types';
+import { getPrincipalsFromRequest } from '../utils';
 
 // Can't throw unauthorized for now, the page will be refreshed if unauthorized
 const generateWorkspacePermissionError = () => {
@@ -358,9 +356,7 @@ export class WorkspaceSavedObjectsClientWrapper {
     const findWithWorkspacePermissionControl = async <T = unknown>(
       options: SavedObjectsFindOptions & Pick<WorkspaceFindOptions, 'permissionModes'>
     ) => {
-      const principals = SavedObjectsPermissionControl.getPrincipalsFromRequest(
-        wrapperOptions.request
-      );
+      const principals = getPrincipalsFromRequest(wrapperOptions.request);
       if (!options.ACLSearchParams) {
         options.ACLSearchParams = {};
       }
